test: extract StringMaker specializations into separate headers
