+++ src/bsoncxx/test/catch.hh
@@ -16,6 +16,8 @@
 
 #include <bsoncxx/v1/detail/macros.hpp>
 
+#include <bsoncxx/test/v1/stdx/optional.hh>
+#include <bsoncxx/test/v1/stdx/string_view.hh>
 #include <bsoncxx/test/v_noabi/to_string.hh>
 
 #include <bsoncxx/document/view_or_value.hpp>
@@ -25,55 +27,12 @@
 #include <bsoncxx/stdx/optional.hpp>
 
 #include <bsoncxx/test/stringify.hh>
+#include <bsoncxx/test/system_error.hh>
 
 #include <catch2/catch_test_macros.hpp> // TEST_CASE, SECTION, CHECK, etc.
 #include <catch2/catch_tostring.hpp>    // Catch::StringMaker
 
-#define THROWS_WITH_CODE_IMPL(_assertion, _expr, _code)                               \
-    if (1) {                                                                          \
-        try {                                                                         \
-            (void)(_expr);                                                            \
-            INFO("expected an exception to be thrown: " #_expr);                      \
-            _assertion(false);                                                        \
-        } catch (Catch::TestFailureException const&) {                                \
-            throw; /* Propagate Catch exceptions. */                                  \
-        } catch (Catch::TestSkipException const&) {                                   \
-            throw; /* Propagate Catch exceptions. */                                  \
-        } catch (std::system_error const& ex) {                                       \
-            using std::make_error_code;                                               \
-            (void)ex; /* Avoid unused variable warnings. */                           \
-            _assertion(ex.code() == (_code));                                         \
-        } catch (...) {                                                               \
-            /* Reuse `*_THROWS_AS` to handle the unexpected exception type. */        \
-            BSONCXX_PRIVATE_CONCAT(_assertion, _THROWS_AS)(throw, std::system_error); \
-        }                                                                             \
-    } else                                                                            \
-        ((void)0)
-
-#define CHECK_THROWS_WITH_CODE(_expr, _code) THROWS_WITH_CODE_IMPL(CHECK, _expr, _code)
-#define REQUIRE_THROWS_WITH_CODE(_expr, _code) THROWS_WITH_CODE_IMPL(REQUIRE, _expr, _code)
-
 namespace Catch {
-
-template <>
-struct StringMaker<std::error_category> {
-    static std::string convert(std::error_category const& value) {
-        return value.name();
-    }
-};
-
-template <>
-struct StringMaker<std::error_condition> {
-    static std::string convert(std::error_condition const& value) {
-        std::string res;
-
-        res += value.category().name();
-        res += ':';
-        res += bsoncxx::test::stringify(value.value());
-
-        return res;
-    }
-};
 
 template <>
 struct StringMaker<bsoncxx::oid> {
@@ -122,24 +81,6 @@ template <>
 struct StringMaker<bsoncxx::types::bson_value::view_or_value> {
     static std::string convert(bsoncxx::types::bson_value::view_or_value const& value) {
         return bsoncxx::test::stringify(value.view());
-    }
-};
-
-template <typename T>
-struct StringMaker<bsoncxx::stdx::optional<T>> {
-    static std::string convert(bsoncxx::stdx::optional<T> const& value) {
-        if (value) {
-            return bsoncxx::test::stringify(value.value());
-        }
-
-        return "{nullopt}";
-    }
-};
-
-template <>
-struct StringMaker<bsoncxx::stdx::optional<bsoncxx::stdx::nullopt_t>> {
-    static std::string convert(bsoncxx::stdx::optional<bsoncxx::stdx::nullopt_t> const&) {
-        return "{nullopt}";
     }
 };
